milldata/serializers.py
from rest_framework import serializers
from .models import Company, Device, Milldata


class CompanySerializer(serializers.ModelSerializer):
    class Meta:
        model = Company
        fields = '__all__'


class DeviceSerializer(serializers.ModelSerializer):
    company = CompanySerializer(read_only=True)
    
    class Meta:
        model = Device
        fields = '__all__'


class MilldataSerializer(serializers.ModelSerializer):
    device = DeviceSerializer(read_only=True)

    class Meta:
        model = Milldata
        fields = '__all__'
